Added HttpStringDownloader, AcceptContentTypes
HttpStringDownloader allows for easy fetching of HTTP resources to be
encoded as strings. Encoding is attempted to be detected automatically,
and falls back to UTF8. An explicit Encoding can be provided as well. An
example:

  new HttpStringDownloader () {
      Uri = new Uri ("http://banshee.fm/foo"),
      AcceptContentTypes = new [] { "text/foo" },
      Finished = d => {
          if (d.State.Success) {
              PresentString (d.Content);
          }
      }
  }.Start ()

On the base HttpDownloader, an AcceptContentTypes property was added
that will cause an exception to be thrown when the HTTP response is
delivered if its Content-Type does not match one of the
AcceptContentTypes. The propery is optinal (an unset AcceptContentTypes
will accept any/all Content-Type responses).

+++ Hyena/Hyena.Downloader/Buffer.cs
@@ -30,7 +30,7 @@ namespace Hyena.Downloader
 {
     public class Buffer
     {
-        public long Length { get; set; }
+        public int Length { get; set; }
         public DateTime TimeStamp { get; set; }
         public byte [] Data { get; set; }
     }
